refactor(dj): gather all packages into one
