+++ .github/workflows/nodejs.yml
@@ -1,7 +1,7 @@
 # This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
 
-name: CI
+name: nodejs
 
 on:
   push:
@@ -42,46 +42,3 @@ jobs:
         with:
           name: lib-${{ matrix.os }}
           path: packages/*/lib/
-
-  test:
-    needs: build
-    strategy:
-      matrix:
-        node-version: [10.x, 11.x, 12.x]
-        os: [ubuntu-latest]
-    runs-on: ubuntu-latest
-    services:
-      redis:
-        image: redis
-        options: >-
-          --health-cmd "redis-cli ping"
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-        ports:
-          - 6379:6379
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v1
-        with:
-          node-version: ${{ matrix.node-version }}
-
-      - id: yarn_cache_dir
-        run: echo "::set-output name=dir::$(yarn cache dir)"
-
-      - uses: actions/cache@v2
-        with:
-          path: ${{ steps.yarn_cache_dir.outputs.dir }}
-          key: ${{ matrix.os }}-${{ matrix.node-version }}-yarn-${{ hashFiles('**/yarn.lock') }}
-          restore-keys: |
-            ${{ matrix.os }}-${{ matrix.node-version }}-yarn-
-
-      - run: yarn && yarn deps
-
-      - uses: actions/download-artifact@v2
-        with:
-          name: lib-${{ matrix.os }}
-
-      - run: yarn test
