--- book.toml
@@ -1,14 +1,0 @@
-[book]
-authors = ["clearloop"]
-language = "en"
-multilingual = false
-src = "book"
-title = "The Darwinia.js Book"
-
-[build]
-build-dir = "docs"
-
-[output.html]
-default-theme = "dark"
-preferred-dark-theme = "ayu"
-git-repository-url = "https://github.com/darwinia-network/darwinia.js"
