--- book/SUMMARY.md
@@ -1,5 +1,0 @@
-# Summary
-
-[The Darwinia.js Book](./README.md)
-[@darwinia/dj](./dj/README.md)
-- [How to test dj locally?](./dj/test.md)
