--- book/guide.md
@@ -1,8 +1,0 @@
-# Guide
-
-## Darwinia
-
-First of all, which darwinia node do you want to connect?
-
-
-
