+++ index.ts
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+import exec from "./src/cmd";
+
+(async () => {
+    exec();
+})();
