--- jest.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-    testRegex: ['/__tests__/.*\\.[jt]sx?$'],
-    preset: 'ts-jest',
-    globals: {
-        'ts-jest': {
-            isolatedModules: true,
-        },
-    },
-    testTimeout: 30000
-}
