--- lerna.json
@@ -1,13 +1,0 @@
-{
-  "packages": [
-    "packages/*"
-  ],
-  "version": "0.1.29-alpha.0",
-  "npmClient": "yarn",
-  "useWorkspaces": true,
-  "command": {
-    "version": {
-      "message": "chore(release): publish %s"
-    }
-  }
-}
