+++ package.json
@@ -1,49 +1,39 @@
 {
-  "private": true,
-  "name": "@darwinia-network/darwinia.js",
-  "version": "0.1.0",
-  "description": "Javascript library for darwinia",
+  "private": false,
+  "name": "@darwinia/api",
+  "version": "0.1.50",
+  "description": "High-level darwinia.js api",
+  "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/darwinia-network/darwinia.js"
   },
   "author": "darwinia-network <hello@darwinia.network>",
   "license": "GPL-3.0",
-  "homepage": "https://github.com/darwinia-network/darwinia.js",
-  "workspaces": ["packages/*"],
+  "main": "lib/index.js",
+  "files": ["lib/**/*"],
+  "dependencies": {
+    "@polkadot/api": "1.30.0-beta.0",
+    "@polkadot/keyring": "3.2.1",
+    "@polkadot/util-crypto": "3.2.1",
+    "axios": "^0.19.2",
+    "yargs": "^15.4.1"
+  },
   "devDependencies": {
     "@commitlint/cli": "^8.3.5",
     "@commitlint/config-conventional": "^8.3.4",
     "@polkadot/types": "1.30.0-beta.0",
-    "@types/ioredis": "^4.16.2",
-    "@types/jest": "^25.2.2",
-    "@types/js-yaml": "^3.12.3",
     "@types/node": "^13.11.1",
     "@types/progress": "^2.0.3",
     "@types/prompts": "^2.0.8",
     "@types/tar": "^4.0.3",
     "@types/yargs": "^15.0.4",
     "husky": "^4.2.5",
-    "jest": "25.0.0",
-    "lerna": "^3.20.2",
-    "ts-jest": "^25.5.1",
     "tslint": "^6.1.1",
     "typescript": "^3.8.3"
   },
   "scripts": {
-    "build": "lerna run build",
-    "clean": "lerna clean",
-    "test": "jest --forceExit --detectOpenHandles",
-    "deps": "lerna exec yarn",
-    "lint": "lerna run lint",
-    "version": "ts-node scripts/version.ts --stable",
-    "publish": "ts-node scripts/publish.ts --stable"
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-      "pre-commit": "git add . && lerna run lint",
-      "pre-push": "lerna run build"
-    }
+    "build": "tsc --strict",
+    "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
   }
 }
