--- packages/api/index.ts
@@ -1,8 +1,0 @@
-import { API, ExResult } from "./src/api";
-import { autoAPI } from "./src/auto";
-import { ShadowAPI } from "./src/shadow";
-import { IEthHeaderThing, BlockWithProof } from "./src/types/block";
-
-export {
-    API, autoAPI, ExResult, ShadowAPI, IEthHeaderThing, BlockWithProof,
-}
