--- packages/api/package.json
@@ -1,26 +1,0 @@
-{
-  "private": false,
-  "name": "@darwinia/api",
-  "version": "0.1.47",
-  "description": "High-level darwinia.js api",
-  "homepage": "https://github.com/darwinia-network/darwinia.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/darwinia-network/darwinia.js"
-  },
-  "author": "darwinia-network <hello@darwinia.network>",
-  "license": "GPL-3.0",
-  "main": "lib/index.js",
-  "files": ["lib/**/*"],
-  "dependencies": {
-    "@darwinia/util": "^0.1.47",
-    "@polkadot/api": "1.30.0-beta.0",
-    "@polkadot/keyring": "3.2.1",
-    "@polkadot/util-crypto": "3.2.1",
-    "axios": "^0.19.2"
-  },
-  "scripts": {
-    "build": "tsc --strict",
-    "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
-  }
-}
