--- packages/dj/package.json
@@ -1,30 +1,0 @@
-{
-  "private": false,
-  "name": "@darwinia/dj",
-  "version": "0.1.47",
-  "description": "High-level darwinia.js api",
-  "homepage": "https://github.com/darwinia-network/darwinia.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/darwinia-network/darwinia.js"
-  },
-  "author": "darwinia-network <hello@darwinia.network>",
-  "license": "GPL-3.0",
-  "main": "lib/index.js",
-  "bin": {
-    "dj": "lib/index.js"
-  },
-  "files": [
-    "lib/**/*"
-  ],
-  "dependencies": {
-    "@darwinia/api": "^0.1.47",
-    "@darwinia/util": "^0.1.47",
-    "yargs": "^15.3.1"
-  },
-  "scripts": {
-    "build": "tsc --strict",
-    "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
-  },
-  "gitHead": "0999fb93cce8f12e1599a2db882148798dcf0c33"
-}
