--- packages/util/index.ts
@@ -1,18 +1,0 @@
-import { Block, IEthBlock, IDarwiniaEthBlock } from "./src/block";
-import { Config, TYPES_URL } from "./src/cfg";
-import { download } from "./src/download";
-import { log } from "./src/log";
-import { whereisPj } from "./src/pj";
-import chalk from "chalk";
-
-// exports
-export {
-    Block,
-    Config,
-    chalk,
-    download,
-    IDarwiniaEthBlock, IEthBlock,
-    log,
-    TYPES_URL,
-    whereisPj,
-}
