--- packages/util/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@darwinia/util",
-  "version": "0.1.47",
-  "description": "Util for developing darwinia javascript library.",
-  "homepage": "https://github.com/darwinia-network/darwinia.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/darwinia-network/darwinia.js"
-  },
-  "author": "darwinia-network <hello@darwinia.network>",
-  "license": "GPL-3.0",
-  "main": "lib/index.js",
-  "files": ["lib/**/*"],
-  "dependencies": {
-    "@polkadot/util": "^3.2.1",
-    "chalk": "^4.0.0",
-    "got": "^10.7.0",
-    "javascript-stringify": "^2.0.1",
-    "progress": "^2.0.3",
-    "prompts": "^2.3.2",
-    "tar": "^6.0.2",
-    "ethereumjs-util": "^7.0.0"
-  },
-  "scripts": {
-    "build": "tsc --strict",
-    "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
-  }
-}
