--- packages/util/tslint.json
@@ -1,23 +1,0 @@
-{
-    "extends": "tslint:recommended",
-    "rules": {
-        "curly": [true, "ignore-same-line"],
-        "max-classes-per-file": false,
-        "max-line-length": {
-            "options": [120]
-        },
-        "new-parens": true,
-        "no-arg": true,
-        "no-bitwise": true,
-        "no-conditional-assignment": true,
-        "no-consecutive-blank-lines": false,
-        "no-console": false
-    },
-    "linterOptions": {
-        "exclude": [
-            "src/json/**/*",
-            "node_modules/**/*",
-            "template/**"
-        ]
-    }
-}
