+++ src/api/index.ts
@@ -1,0 +1,8 @@
+import { API, ExResult } from "./api";
+import { autoAPI } from "./auto";
+import { ShadowAPI } from "./shadow";
+import { IEthHeaderThing, BlockWithProof } from "./types/block";
+
+export {
+    API, autoAPI, ExResult, ShadowAPI, IEthHeaderThing, BlockWithProof,
+}
