+++ src/util/index.ts
@@ -1,0 +1,18 @@
+import { Block, IEthBlock, IDarwiniaEthBlock } from "./block";
+import { Config, TYPES_URL } from "./cfg";
+import { download } from "./download";
+import { log } from "./log";
+import { whereisPj } from "./pj";
+import chalk from "chalk";
+
+// exports
+export {
+    Block,
+    Config,
+    chalk,
+    download,
+    IDarwiniaEthBlock, IEthBlock,
+    log,
+    TYPES_URL,
+    whereisPj,
+}
