+++ src/util/static/config.json
